monitoring (fix): Change variable used in ansible for monitoring
Taking into consideration commit: 12492da7486784a8282c796988a96befcfbdaf5c

According to ansible, the variable to use was 'groups' and not
'group_names':
"group_names is a list (array) of all the groups the current host is in. "
"groups is a list of all the groups (and hosts) in the inventory."

This is explained because we want to know globally if 'monitoring' is
present in the ansible host file, and not for a specific node.