module: Initial commit
This is the first commit of a python's file inspired of the check done
in python. This tends to start answering the problems that will be faced
while trying to solve: https://github.com/redhat-cip/rcip-openshift-ansible/issues/19#issuecomment-194915476

Note: this is still a work in progress. There are many things missing,
but it should be a good start to continue from.